+++ lib/chef/knife/bootstrap_windows_base.rb
@@ -278,6 +278,14 @@ class Chef
           warn_chef_config_secret_key
         end
 
+        bootstrap_architecture = Chef::Config[:knife][:bootstrap_architecture]
+        if bootstrap_architecture && ![:x86_64, :i386].include?(bootstrap_architecture.to_sym)
+          raise "Valid values for the knife config :bootstrap_architecture are i386 or x86_64.  Supplied value is #{bootstrap_architecture}"
+        end
+        if Chef::Config[:knife][:architecture]
+          raise "Do not set :architecture in your knife config, use :bootstrap_architecture."
+        end
+
         validate_name_args!
 
         # adding respond_to? so this works with pre 12.4 chef clients
@@ -319,6 +327,9 @@ class Chef
         end
 
         wait_for_remote_response( config[:auth_timeout].to_i )
+
+        set_target_architecture(bootstrap_architecture)
+
         ui.info("Bootstrapping Chef on #{ui.color(@node_name, :bold)}")
         # create a bootstrap.bat file on the node
         # we have to run the remote commands in 2047 char chunks
@@ -409,6 +420,29 @@ behavior will be removed and any 'encrypted_data_bag_secret' entries in
 'knife.rb' will be ignored completely.
         WARNING
         ui.info "* " * 40
+      end
+
+      # We allow the user to specify the desired architecture of Chef to install or we default
+      # to whatever the target system is.  We assume that we are only bootstrapping 1 node at a time
+      # so we don't need to worry about multipe responses from this command.
+      def set_target_architecture(bootstrap_architecture)
+        session_results = relay_winrm_command("echo %PROCESSOR_ARCHITECTURE%")
+        if session_results.empty? || session_results[0].stdout.strip.empty?
+          raise "Response to 'echo %PROCESSOR_ARCHITECTURE%' command was invalid: #{session_results}"
+        end
+        current_architecture = session_results[0].stdout.strip == "X86" ? :i386 : :x86_64
+
+        if bootstrap_architecture.nil?
+          architecture = current_architecture
+        elsif bootstrap_architecture == :x86_64 && current_architecture == :i386
+          raise "You specified bootstrap_architecture as x86_64 but the target machine is i386.  A 64 bit program cannot run on a 32 bit machine."
+        else
+          architecture = bootstrap_architecture
+        end
+
+        # The windows install script wants i686, not i386
+        architecture = :i686 if architecture == :i386
+        Chef::Config[:knife][:architecture] = architecture
       end
     end
   end